refactor: split core type modules
