+++ src/rust/core/src/model/mod.rs
@@ -1,0 +1,10 @@
+mod output_type;
+mod result;
+mod settings;
+
+#[cfg(test)]
+mod tests;
+
+pub use output_type::OutputType;
+pub use result::{GeneratorResult, HttpFile};
+pub use settings::GeneratorSettings;
