+++ src/rust/core/src/model/output_type.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
+pub enum OutputType {
+    #[default]
+    OneRequestPerFile,
+    OneFile,
+    OneFilePerTag,
+}
