+++ src/rust/core/src/normalized/mod.rs
@@ -1,0 +1,17 @@
+mod document;
+mod http;
+mod parameter;
+mod request_body;
+mod schema;
+
+#[cfg(test)]
+mod tests;
+
+pub use document::{
+    NormalizedOpenApiDocument, NormalizedOperation, NormalizedServer,
+    NormalizedSpecificationVersion,
+};
+pub use http::{NormalizedHttpMethod, NormalizedParameterLocation};
+pub use parameter::{NormalizedInlineParameter, NormalizedParameter};
+pub use request_body::{NormalizedInlineRequestBody, NormalizedMediaType, NormalizedRequestBody};
+pub use schema::{NormalizedSchema, NormalizedSchemaProperty, NormalizedSchemaType};
